feat(ci): generate ci-dispatch-manifest via standalone tsx (skip Astro build)
Adds a standalone tsx generator that reads project MDX frontmatter and
emits .github/ci-dispatch-manifest.json byte-identically to the Astro
build path, without paying for a full Astro build.

- Extract shared buildManifest() + entry types into
  src/data/ci/manifest-builder.ts (no astro:content dep).
- Extract ProjectSchemaWithEngine into src/data/ci/project-schema.ts
  (astro/zod, same instance -> no cross-boundary nested-object strip);
  content.config.ts imports it instead of defining inline.
- ci-registry.json.ts route slims to getCollection -> buildManifest.
- scripts/gen-ci-manifest.mts: glob + gray-matter + schema.parse ->
  buildManifest; write or --check. Schema parse reproduces Astro key
  order + unknown-field stripping, so output is byte-identical.
- nx targets gen:ci-manifest and gen:ci-manifest:check.
- ci-manifest-guard.yml and utils-update-version-toml.yml regenerate
  via gen:ci-manifest instead of astro build + sync:ci-manifest.
